topology: pre-process-class: add function to look up token_ref for an attribute in class
Some attributes may have the token_ref set which is
used to look up the token value for the tuple data
that is appended to the object's private data.

For example, in the buffer widget object:
	DefineAttribute."size" {
		# Token reference and type
		token_ref	"sof_tkn_buffer.word"
	}

The token_ref must include the reference to the vendor
token object name followed by the type of the tuple.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>