topology: pre-processor: Fix pre-processor defs parsing
Using snd_config_merge() with tplg_pp->define_cfg() deletes the command
line definitions after the first merge. So use a temporary config
instead so that when recursively merging definitions, the command line
variables can always override the defaults.

Also, to allow for the case where there are no default definitions in
the conf file but there are command line definitions, do not return an
error but create an empty Define subtree so it can be merged with the
command line definitions.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>