topology: pre-processor: Make instance attribute mandatory for all classes
Make in mandatory for all class definitions to include the instance
attribute. This will ensure that the objects instantiation is uniform
irrespective of the type of object. For example, a widget can be
instantiated as

Object.Widget.gain.1 {}

And a pipeline can be instantiated as

Object.Pipeline.volume-capture.1 {}

Additionally, when including conf files with its own object
instantiations, this will allow for the alsatplg compiler pre-processor
to uniquify the instance IDs for that there are no clashes. This is done
in the next patch.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>