topology: pre-processor: Uniquify include object instances
Included conf files could contain additional object instantiations with
the same instance ID as the top-level conf file. For example, the
top-level conf file could contain a PCM object as below:
Object.PCM {
    pcm.10 {
		name    "ssp-capture"
    }
}

If the included conf file could contain a PCM object with the same
instance ID of 1 like:
Object.PCM {
        pcm.10 {
                name    "DMIC"
        }
}

In this case when merging the above conf file with the top-level file,
the first PCM object with name "ssp-capture" will get overridden with the
new one "DMIC". To prevent this and add 2  separate PCM instance, this
function makes sure that the instance ID for the PCM object in the
included file does not clash with the top-level object instances. So the
new conf node will modified to automatically increment the instance ID
to be larger than the largest ID in the top-level conf as below before
getting merged.

Object.PCM {
        pcm.11 {
                name    "DMIC"
        }
}

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>